parser: Handle columns with CustomExtractor
Currently if column is using a CustomExtractor the kind is
set to 'string'. This means we won't be able to use columns
values even if the underlying type in int/float. This commit
handles the cases if underlying type in int64/float64. I need
this for dns latency field (time.Duration)

Signed-off-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>