cmd: Permit deploying bcc flavor.
This commit adds the "--bcc" CLI flag to deploy the bcc flavor of Inspektor
Gadget.
It cannot be used with "--image", as "--bcc" will use "latest-bcc".

This commit also removes CAP_SYS_MODULE from deploy.yaml as default does not
need it.
Since bcc gadgets need it, we add it during deployement when bcc flavor is
detected.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>