gadget/run: client shouldn't pull image and parse BTF
This is hugely experimental commit.

The client still pulls the image and parses the BTF but the informations
are not used by the client at all. (Still need to cut off that part)

The server sends the dynamic fields from the BTF struct encoded in JSON
through gRPC to the client.
There the client now knows which columns it is expecting. It creates a
virtual column for every dynamic field.

The server always sends the Event itself in JSON through gRPC again (the
JSON derived from the columns library, see implementation). The client
decodes the JSON and unpacks it into the runType.Event struct. For the
dynamic fields it uses the Data field of runType.Event.
For that the Data field gets used as a map[string]string, where the map
key is the dynamic column name and the map value is the value for that
column.

Signed-off-by: Burak Ok <burakok@microsoft.com>