experimental: Only init operators when they are really needed by the gadget
It's not possible to determine which operators are needed by a
containerized gadget before initializing it. This commit modifies the
kubeipresolver and the run gadget to use an extra interface that tells
if the gadget actually has endpoints.

One issue is that this is somehow duplicated with the CanOperate() logic.