Dockerfiles: Add bpftrace to default image.
When we first added bpftrace, i.e. script gadget, to Inspektor Gadget, we only
focused on the default image and not the CO-RE one in 36e2e489464e
("Dockerfiles: Add bpftrace binary to default image").
As, we just make CO-RE the default, we need to add bpftrace here to not loose
features.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>