gadgets/run: Support eBPF parameters
Gadgets usually define different constants in the eBPF code to change
its behaviour. This commit implements the logic allow users setting
these flags from the CLI.

The gadget information to carry the different parameters exposed by the
bpf programs. This is gathered from the list provided in the gadget
definition and by using the BTF information of the gadget.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>