ig: add daemon mode
This adds a daemon mode to ig; ig will listen on either a unix or tcp socket
for incoming gRPC requests that it will then execute. This allows using ig as
a service rather than starting it individually for every gadget. A new binary
will be added to be able to serve as a client.

Signed-off-by: Michael Friese <mfriese@microsoft.com>