CI: Automatically build and push gadgets
If the event is a PR then the image is pushed to ttl.sh, so we can run
integration tests. If the event is different, the image is pushed to
the ghcr.io/inspektor-gadget/ repository.

The integration tests are update to use the OCI images instead of the
objects too.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>