cmd/common: expose AddFlags and remove registering global runtime flags
Global runtime flags have previously not been used but they will be used by gadgetctl for now.
However, gadgetctl will register those flags on the root command level and not per-gadget.
This change exports AddFlags so that it is available to the main packages and removes
registering it in AddCommandFromRegistry to avoid duplicate params.

Signed-off-by: Michael Friese <mfriese@microsoft.com>