Refactor pkg/oci
This is a complete refactoring of the oci package. The whole idea is to
implement all the OCI related functionalities we need in this package,
avoiding other parts of the code having to interact with packages like
"github.com/docker" or "oras.land/oras-go/v2".

The reasoning behind is:
- We can concentrate all OCI related code in the same place
- It's a lot easier to test if it's a separated package

This package now provides methods for pulling, pushing, tagging and
building OCI images. Other methods that we previously exposed are
now internal.

This commit also:
- Pass context object everywhere
- Uses -ing in all error messages
- Small esthetic changes to cmd/*

NOTE: Ideally this should be multiple commits, but it's rather difficult
to split a refactor like it into smaller pieces.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>