gadgettracermanager: add option to run service on a tcp socket and make it the default
This change renames parameter `service-socketfile` to `service-host` and now allows to specify
either a unix socket or a tcp address to listen on. It uses the same format we use for `ig -H`
(so either unix:///path/to/unix/socket or tcp://127.0.0.1:8080). The default is to use tcp
now that we're not running in the host network namespace anymore, as this allows to use
port forwarding via the Kubernetes API server instead of having to execute `socat` anymore.

Signed-off-by: Michael Friese <mfriese@microsoft.com>