Avoid decoding BTF on the client
This commit changes the logic to avoid the client having to decode BTF
information. The whole idea consists on having the "columns descriptors"
in the GadgetInfo structure. These descriptors are generated by the
server and used by the client to decode the data it receives.

The data is sent as a binary blob [][]byte. The first array is the bpf
event, the second one are fixed type fields and the other ones are used
for strings. This commit adds some helper methods to add fields to the
event.

This commit is an intermediate step in order to enable support for wasm.
The next work is to start using it from operators to decouple them
from the tracer implementation.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>