operators: Add CanOperateOnContainerizedGadget()
built-in gadgets use a gadget descriptor to expose the features they
support (like filtering, enriching, etc.). This approach is not valid
for containerized gadgets as the features implemented by each gadget
could be different. This commit adds a new CanOperateOnContainerizedGadget()
function uses the containerized gadget features to check if a given
operator support that gadget.

This commit improves the efficiency as not all operators aren't enabled
for a containerized gadget.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>