cmd: Ignore unknown flags when parsing params early
When connecting to a server we need to have information about the
connection itself. That is why we parse the flags as early as possible.
Even before we have all of them available, because the flags are derived
from the server's information.

This commit enables the option to ignore unknown flags on this early
flag parsing, so we can get information on the connection flags without
erroring out.

Fixes: e89e6828c334 ("add gadgetctl as client for ig")
Fixes: cd5768eb8078 ("pkg/runtime/grpc: add option to connect to the gadget
                      service using either tcp or unix socket")

Co-authored-by: Michael Friese <mfriese@microsoft.com>
Signed-off-by: Burak Ok <burakok@microsoft.com>