docs: crd: disable publication
The CRD reference documentation is available at:

https://github.com/inspektor-gadget/inspektor-gadget/blob/main/docs/crds/traces.gadget.kinvolk.io.md
https://inspektor-gadget.io/docs/latest/crds/traces.gadget.kinvolk.io/

However, Hugo does not generate the page correctly because it omits raw
HTML and replaces it with the following comment:

    <!-- raw HTML omitted -->

Given that the CRD is soon to be deprecated, it is not worth fixing it.
This patch disables the publication of the web page.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>