run: Support multiple socket filter programs
The logic to create the network tracer (needed by socket filter programs)
was placed in NewInstance(), unfortunately it has a lot of drawbacks:
1. NewInstance() is called many times in the workflow without running
the gadget, it was causing the network tracer to be instantiated
multiple times without any purpose.
2. Creating the network tracer requires root. It was making `ig --help`
to print some annoying warnings (https://github.com/inspektor-gadget/inspektor-gadget/issues/2108)
3. The network tracer was always created, even for gadgets that don't
need it, also only a single socket filter program was supported.

The main motivation to have this logic on NewInstance() was that the
network tracer has to be created before calling AttachContainer(). This
commit moves all that logic to Init() (that is called before
AttachContainer()) and enables the support for multiple socket filter
programs on a gadget.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>