run: Fix lot of leakages
The GadgetStartStop interface was removed a long time ago in a3e5871a338a ("pkg/runtime/local: Move context and timeout management into tracers"),
so the Stop() method was never being called. It was causing to leak
all BPF objects of the gadgets being run.

This commit also fixes a leakage on the network tracer as it was never
being closed.

Fixes: 07ca0fb4e97a ("Initial support for containerized gadgets")
Fixes: c1cc966f4f16 ("implement networking gadgets for containerized gadgets")

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>