cmd/deploy log level in ig-k8s pod (#2312)
Add a new command line option to ig-k8s to set the log level via --daemon-log-level.

This is useful for debugging issues in the field. 
The new command will set the environment variable **GADGET_TRACER_MANAGER_LOG_LEVEL** in the ig-k8s pod. The default value will be **info**.

Changes were made to the daemonset chart to add the new command line option and to set the environment variable in the pod.
Changes were made to the gadget container to set the log level based on the environment variable.
The log level is set in entrypoint.go, gadgettracermanager/main.go and cleanup.go.

To set the GADGET_TRACER_MANAGER_LOG_LEVEL to trace, run the following command:

```shell
 ./kubectl-gadget deploy --daemon-log-level=trace
 ```

The log level can be set to any of the following values from the logrus package:
- trace
- debug
- info
- warn
- error
- fatal
- panic

Signed-off-by: Glenn Hinks <ghinks@yahoo.com>