pkg/run: Set gadget_heap map size to size of event tracer.
In case ring buffer are not available, perf buffers will be used and gadget_heap
will be created.
By default, gadget_heap size is set to MAX_EVENT_SIZE in buffer.h which default
value is 10240.

In this commit, we compute the event tracer structure size and set gadget_heap
accordingly.
This permits reducing memory footprint, as we only use what we really need.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>