gadgets: Add tchandler package
The tchandler packet handles how TC programs (only SchedCLS for now)
are attached to containers and network interfaces.

The logic is similar to the existing pkg/gadgets/internal/networktracer
for socket programs, however there are some differences:

1. It's possible to attach TC programs in ingress or egress, hence
the "sense" of attachment has to be passed around.
2. This package is designed only for image-based gadgets, hence it
implement support for polling events from the ebpf program.

It could be possible to unify these two components, but it's not so easy
to do right now, hence this is future work once the built-in gadgets
are removed.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>