ci: Add dedicated job to run CodeQL analysis for ig.
With the current job, we build ig using docker.
As a consequence, CodeQL does not detect the binary was built and will try to
build it itself, which adds extra time in the CI pipeline [1].
So, we add a dedicated job to build ig with go build and then run CodeQL
analysis here.

Fixes: f04d95b078d5 ("ci: Add CWE checks for ig.")
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://github.com/github/codeql/issues/15405#issuecomment-1905608507