trace/dns + trace/exec + trace/open: fix unsafe typecasting
Before this change, we cast incoming events from the RawSample byte slice directly to expected types (like in other gadgets).
However, in these gadgets, we optimized the data we send in eBPF to reduce load by not sending empty fields. Casting those
abbreviated events to the full struct though makes them not backed by memory anymore - and thus unsafe to use.
This change adds a new type for each gadget containing only the guaranteed parts that are sent and fills the (dynamic) remainder
directly from the RawSample data. In case of changes to the original event layout, these new structs have to be updated as well.

Fixes #2337.

Signed-off-by: Michael Friese <mfriese@microsoft.com>