wip: unify ig and ig-k8s tests
Currently we have a lot of duplication between the ig-k8s/kubectl-gadget
(integration/inspektor-gadget/) and ig in k8s (integration/ig/k8s/) tests.

This commit is a POC showing how we can unify those tests. The idea is
to have a single test file and by using flags to indicate if this has
to run kubectl-gadget (for ig-k8s) or ig.

The tests cases should be almost the same, however it's possible to
customize some of the things by checking DefaultTestComponent inside
each test.