integration: Disable installing Security Profile Operator
The security profile operator is only used by TestAuditSeccomp to
install a simple seccomp profile in the cluster. This is an overkill.

This commit uses a simpler approach by copying the profile by using
a DaemonSet to the node and disables installing this by default.
The code to deploy/undeploy the SPO isn't removed as it could be useful
to test the "advise seccomp" gadget in the future.

Some reasons to avoid deploying the SPO are:
- It needs quite a lot of resources
- It takes a lot of time to be deployed
- Some times it fails to be deployed, and also to be cleanup, leaving
  it running on the cluster

Signed-off-by: Mauricio Vasquez <mauriciov@microsoft.com>