pkg/run: Don't rely on GADGET_TRACER as it is optional
GetTracerInfo gets the tracer information from what was defined by
the GADGET_TRACER macro. However, such macro is optional, so we
must not rely on it. Additionally, by the time
handleTracerMapDefinition() is called, the metadata was already
built (either from the metadata file or from the eBPF code), so we
can get the tracer information from there.

Signed-off-by: Jose Blanquicet <josebl@microsoft.com>