run: Print fields on same order as defined in the metadata
In some cases it's not possible to order the fields in the event
structure in a human-friendly way (padding and other reasons), hence
it's possible to change the order of those fields in the metadata file.
This commit updates the logic to use the same order defined in such
file instaed of relying on the event structure layout.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>