pkg/gadgets: Fixes trace bind for kernel 6.6 and younger.
In kernel 6.6, the bitfields in struct inet_sock where moved to a dedicated flag
[1, 2, 3, 4].
To adapt to this change, a CO-RE solution was proposed in iovisor/bcc for
bindsnoop [5].
This is a good solution, as it will work on kernel backporting the patches.

This commit adapts the bindsnoop solution to Inspektor Gadget.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://lore.kernel.org/all/20230816081547.1272409-1-edumazet@google.com/
[2]: https://github.com/torvalds/linux/commit/3f7e753206bb
[3]: https://github.com/torvalds/linux/commit/ca571e2eb7eb
[4]: https://github.com/torvalds/linux/commit/4bd0623f04ee
[5]: https://github.com/iovisor/bcc/pull/4846/commits/a55b5a0c920f