image/build: Generate BTF files with btfgen
btfgen[0] is used to generate BTF files that are needed to run an eBPF
program in a kernel that doesn't provide BTF information. It's possible
that gadget developers want to run their gadgets in those situations,
hence this commit implements the logic to invoke btfgen and store the
generated BTF files in a new layer on the gadget image.

Next commit will implement the logic to use those files.

[0]: https://kinvolk.io/blog/2022/03/btfgen-one-step-closer-to-truly-portable-ebpf-programs/

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>