gadgettracermanager: Explictly import kubemanager operator
Currently, the kubemanager operator is registered (its init() is called)
because it's imported by the kubenameresolver operator, which is
imported by the gadget-collection trace network gadget. It means that if
we decide to remove such a gadget, the kubemanager operator won't be
registered anymore.

This commit adds the kubemanager operator to the gadgettracermanager to
avoid dependence on other operators or gadgets.

Signed-off-by: Jose Blanquicet <josebl@microsoft.com>