pkg/oci: make fixMetadataOwner more generic
fixMetadataOwner() was previously added for the following reason:

'ig image build --update-metadata' overwrites gadget.yaml but since ig
runs as root, the file gadget.yaml initially ends up belonging to root.
The function fixMetadataOwner() fixes this with a chown.

In order to reuse fixMetadataOwner() for a different file, this patch
makes it more generic.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>