image list: add creation date
The OCI image spec defines the creation date annotation:
https://github.com/opencontainers/image-spec/blob/main/annotations.md

org.opencontainers.image.created date and time on which the image was
built, conforming to RFC 3339.

This patch ensures that the annotation is set with the same value for
both amd64 and arm64.

Then, the date is printed in the image list command:

```
$ sudo -E ig image list
INFO[0000] Experimental features enabled
REPOSITORY  TAG     DIGEST       CREATED
trace_dns   latest  d98eea3ede9e 4 minutes ago
```

Note: the creation date will also be useful for the future Artifact Hub patch
because the 'createdAt' field is mandatory in artifacthub-pkg.yml:

https://github.com/artifacthub/hub/blob/master/docs/metadata/artifacthub-pkg.yml
```
createdAt: The date this package was created (RFC3339 layout) (required)
```

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>