README.md: update Helm charts badge
Update the label to: "Helm charts"

Update the link to the URL of the Helm chart instead of the search
webpage.

Note that the URL still contains "gadget". We are looking to change it
to "inspektor-gadget" but at the moment it is linked to the chart name
and renaming it would break the updates for users.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>