eks-cleanup: Don't error when stack can't be deleted
This is probably the case when the VPCs are not deleted yet. Swapping
the order to delete VPCs before the Cloudformation stacks is also not
feasible since VPCs can't be deleted when the stack is still running.

So we just try to delete both and in the next run the corresponding
stack or VPC will be deleted

Signed-off-by: Burak Ok <burakok@microsoft.com>