image/build: Use correct Inspektor Gadget code in in-tree gadgets
Building in-tree gadgets is a bit more complicated as we want
them to use the current files to be sure they take any
changes we're working on. (It's specially relevant
for the api exposed to gadgets in pkg/operators/wasm/api/)

One big challenge is that (ideally) the build process should continue
to work when using a container or when building locally on the host (--local).

This commit achieves it by using a replace directive with a relative
path to the inspektor-gadget folder. The location of the Inspektor Gadget
source code can be specified by defining the IG_SOURCE_PATH env variable,
it causes the build command to mount it into the builder container.

This commit isn't totally clean, however other solutions seemed even more complicated:
- Copy the inspektor gadget source code to the builder image
 - pros:
  - We don't need to mount anything on the container
 - cons:
  - The ebpf builder image will be implicitly associated with an inspektor gadget version
- Always use --local for building in-tree gadgets
 - pros:
  - We don't need to worry about this issue at all
 - cons:
  - Complicates the CI as we'll need to install all dependencies locally to compile gadgets

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>