ci: Skip signing ig-k8s image/gadget images on fork
Currently, in order to publish ig-k8s image / gadget images on
a fork users will have to:
- Set secrets for cosign (COSIGN_PRIVATE_KEY/COSIGN_PASSWORD)
- Update the public key in source (pkg/resources/inspektor-gadget.pub)
the idea of this change is to improve this workflow by not signing images
by default on forks. We won't depend on public-key-check when
publishing images, but add it in the release job to ensure images
are published with correct keys.

Signed-off-by: Qasim Sarfraz <qasimsarfraz@microsoft.com>