testing: implement k8s runtime
The k8s runtime runs containers using k8s. The interface exposed
is the same as runtimes as docker and containerd. Even if Kubernetes
and container runtimes aren't the same concept, for the scope of testing
Inspektor Gadget we can use then in the same way: to run containers.

How to use:

$ cd /gadgets/trace_open/test

--> Test ig locally with docker
$ IG_PATH=ig IG_RUNTIME=docker go test -exec "sudo -E" .

--> Test with kubectl gadget and k8s (notice no sudo is needed in this case)
$ IG_PATH=kubectl-gadget IG_RUNTIME=kubernetes go test .

--> Test ig with kubernetes
$ IG_PATH=ig IG_RUNTIME=kubernetes go test -exec "sudo -E" .
TODO: this doesn't work for me...