pkg/operators: Fix tracing host events for mount namespace ID attacher gadgets.
Some attacher-based gadgets, like traceloop and advise seccomp-profile, need
the mount namespace ID of the container to bet set.
With --host, we only created a fake container structure containing only PID 1.
So, traceloop and advise seccomp-profile would not work correctly as their
required information were not set.

To cope with this, we extend the fake container structure to also contain the
mount namespace ID of PID 1.

Fixes: 740976d3ab0f ("pkg/operators: Enable tracing hosts events for attacher gadgets.")
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>