formatters: Avoid changing name of the field
The current logic of the formatters changes the name of the original field
(adding a "_raw" suffix) and creates a new field with the original name.

This behavior is not right as the name and type of the fields depends
on the operators being run, hence, it could cause issues when parsing the
output of Inspektor Gadget and enabling / disabling the formatters
operator.

This commit fixes it by keeping the original name in all the cases. The
new name is derived from an annotation on the metadata file or by
removing the _raw suffix from the original name.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>