cmd/ig: Preprend host.HostRoot to runtime socket path.
In list-containers, we create an ig manager with NewManager()
instead of Init() [1].
In this specific case, the runtimes were configured with
utils.AddCommonFlags() [2].
Sadly, this function does not handle appending of host root to
runtime socket paths.

This commit fixes this behavior by appending host.HostRoot to
runtime socket paths.
This does not change the default behavior, as host.HostRoot equals
"/", but it will make list-containers taking into account HOST_ROOT
value.
This can be useful when ig is run from a container.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://github.com/inspektor-gadget/inspektor-gadget/blob/902780751407/cmd/ig/containers/containers.go#L52
[2]: https://github.com/inspektor-gadget/inspektor-gadget/blob/902780751407/cmd/ig/utils/flags.go#L87-L95