testing: Consolidate match logic in pkg/testing/match
This commit does different things:
1. Move logic from helpers.go to match.go in order
to have all the matching logic in the same place.
2. Consolidate the testing functions into two:
MatchEntries() and MatchAllEntries() by creating a
new OutputMode type that defines how to parse the
output. It (hopefully) makes the code more clear
as there aren't a lot of different functions
around. This commit also adds some documentation
explaning what those different modes are for.
3. Update all tests to use the new functions.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>