tests: Disable TestTraceNetwork on cri-o
This test is failing a lot (see issue #2358), creating a lot of noise
and forcing us to rerun the CI multiple times.

It's not ideal to disable a failing test, but it's better than wasting
a lot of time rerunning the CI. We'll keep the issue for fixing it
once we have time.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>