ebpf/formatters: Support formatting bitfields
This commit adds support for providing a human readable version of a
bitfield.

This is done by using the same logic for enums, in this case, the gadget
author has to define an enum with all values of the flags and that its
name has the "_set" suffix, with this information the formatter will
recognize and format it on the right way.

Developers can use the "ebpf.formatter.bitfield.separator" annotation to
set the separator to be used when joining the flags, by default "|" is
used.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>