datasource: Implement columns replace annotation
Add a way to replace the content of a field when
formatting it as a column.

This support is useful for printing the pretty
version of complex types that can't be printed as
a column, for instance, the json representation of
an L4 endpoint is something like:

{
  "addr": "1.1.1.1",
  "port": 443,
  "proto": 6,
  "version": 4
}

The formatter could create a new virtual columns
with a pretty representation "addr:port" and
indicate that it has to be printed when this field
is required.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>