gadgets/trace_dns: Parse DNS packet using wasm
This commit is very similar to d772f2ef04c7 ("dns: Parse DNS packet in user space"),
but in this case the parsing is done in wasm because it's an image-based
gadget.

Unfortunately it seems that using gopacket with tinygo isn't possible,
so I decided to use golang.org/x/net/dns/dnsmessage that provides
the functionality we need.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>