gadgets: Use annotations instead of attributes
This commit moves the fields configuration from the struct to a new
datasources section in the metadata file. This is done to allow the user
to specify the configuration for fields that are not strictly related to
ebpf, for instance, fields addded by operators like wasm and the formatter.

Also, this commit only uses annotations instead of attributes, as it's
more flexible. (This was a leftover from the previous refactoring).

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>