pkg/operator/otel-metrics: add support for metrics from eBPF and rendering to datasource
This adds several changes to the otel metrics operator: it can now work with a global or a
local metrics provider. The former is used, when the user wants to export metrics to the
Prometheus endpoint, the latter for everything else.

Additionally, given the metrics.print annotation is set on a datasource, metrics will be
rendered in human-readable form to a newly created datasource (that then can be printed
to console using the CLI operator.)

I've also changed the initialization so that resources needed for the local metric providers
will only be allocated when the gadget is run and freed when it is stopped.

Signed-off-by: Michael Friese <mfriese@microsoft.com>