ebpf: Support zero length arrays fields
Zero length arrays are useful when the structure is the header of a
variable-length object [0]. This commit extends Inspektor Gadget to
support them. The whole idea is to register the structure as static
fields, and if a zero length array is detected, it's registered as a
bytes array field.

TODO:
- only tracers are supported, should it be more generic?
- Logic is implemented inside the bpf operator, would it make sense to
have it directly on the datasource package?

[0]: https://gcc.gnu.org/onlinedocs/gcc/Zero-Length.html

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>