docs: Add gadgets documentation
Add documentation for trace_exec and trace_open based on the existing one.
This commit creates a new README.mdx file that allows to use Tabs to
render the documentation in a better way.

A minimal README.md file is kept on each gadget to be able to show it
on artifacthub.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>