pkg/runtime/grpc: prevent datasource mix-up
Before this change it could happen that dsMap got mixed-up content: dsNameMap["nonexistent"] would result in "0",
which is usually a valid datasource - this is also true for locally created datasources, as they're not added to
the dsNameMap, but still get created before the iteration here in LoadGadgetInfo(). This adds a check so that only
datasources present at the remote side will be added to the dsMap.

Signed-off-by: Michael Friese <mfriese@microsoft.com>