pkg/gadgets: Fix tree output for snapshot process.
With the changes added in d61e5ef270f1
("cc: Move BasicK8sMetadata to pkg/types and use it in CommonData") this output
mode no more worked in non k8s context as there would be an empty string for
K8S.ContainerName.
Let's use the mount namespace ID to group processes in a container instead.

Fixes: d61e5ef270f1 ("cc: Move BasicK8sMetadata to pkg/types and use it in CommonData")
Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>