fsnotify: add new fields
New fields:
- ppid
- uid
- gid
- pcomm

For both the tracer and the tracee processes.

This introduces a new 'struct process' where the common fields can be
located. This avoids duplicating most of the fields in two copies for
the tracer and the tracee. The field mntns_id and gadget_uid/gadget_gid
are not located in 'struct process' due to current limitations in
Inspektor Gadget.

The default mntns_id for container enrichment and filtering is the
one on the tracee process because our most common use case is to get
container enrichment for the tracee process.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>