build: Include layer with help information
Gadgets are distributed using OCI images and are designed to be run with
the ig or kubectl-gadget commands. However, users could be confused and
try to run those images with docker, which will produce a difficult to
understand error message.

This commit adds a layer to the OCI image with a very small binary that
prints a message to the terminal telling the users how to run the gadget.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>