gadgets/top_tcp: Mark received and sent fields as metric counters
Given that the `sent` and `received` fields act already as counters,
let's use the `gadget_counter__u32` type to make Inspektor Gadget manage
them as metric counters. Doing so, we simplify the users' life because
they will just need to select the key to make Inspektor Gadget start
exporting the metrics for the `tcp` datasource.

Signed-off-by: Jose Blanquicet <josebl@microsoft.com>