gadgets: Automatically add annotations based on the field types
Gadgets use a lot of well-known fields like pid, tid, uid. etc. This
commit adds logic in Inspektor Gadget to automatically add annotations
for those types to avoid having to repeat the same thing in the
metadata file for all gadgets. It is still possible for gadget
developers to override the defaults.

This commit also moves the annotation logic to a more neutral way:
- annotations names to pkg/metadata/v1/annotations.go
- annotations templates to pkg/types/annotations.go

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>