cmd/common/oci: validate/reuse node information when attaching to existing gadgets
If a gadget instance has nodes defined, make sure that - when trying to attach - those nodes are being
used, as only those can extract the GadgetInfo. If when attaching nodes are specified as well, make sure
that those are in the set of nodes that have been specified for the gadget instance.

Signed-off-by: Michael Friese <mfriese@microsoft.com>