container-hook: additional coherence checks
Add coherence checks:
- the pid file must not exist before the 'runc create' command.
- the container actually created a new mount namespace.

At the time when the 'runc create' command starts, it is still in the
host mount namespace. This patch records the mount namespace to compare
it with the mount namespace of the container process. If it is the same,
the 'runc create' command didn't actually create a new container.

This reduces the risk of users executing invalid 'runc create' commands
that interfere with Inspektor Gadget.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>