operators/wasm: create dedicated context instead of relying on gadgetCtx
Background: the context used will also be propagated to any callbacks registered in
any of the lifecycle hooks. So whenever a gadgetCtx is cancelled, this will effectively
also stop any callbacks to be made anymore. In case we want to clean up things in WASM,
we however need a valid context to be still alive until we reach Stop(), where the
dedicated context will finally be closed as well.

Signed-off-by: Michael Friese <mfriese@microsoft.com>