operators/ebpf: Enable collecting statistics for gadgets
Implement logic to collect statistics for gadgets and ebpf programs
in general.

The logic creates a new datasource that emits the runcount, runtime,
map memory and map count for gadgets and ebpf programs. An entry for
each ebpf program in the system is created and an entry summarizing the
gadget is added as well.

TODO: How to distinguish the bpfstats gadget in a better way?

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>