operators/ebpf: Enable collecting statistics for gadgets
This commit implements logic to collect statistics for gadgets and ebpf
programs.

This is done by two parts:
1. The existing eBPF operator is extended to keep track of the eBPF
programs and maps created when loading a gadget. This is used to stablish
the gadget to program relationship.
2. A new datasource operator that emits the metrics based on the
information collected above is implemented.

This logic can work on two different ways according to the programs param:
1. per program statistics: It provides granular statistics for each
ebpf program running on the system, even for ebpf programs not controlled
by Inspektor Gadget. Inspektor Gadget controlled programs are enriched
with the gadget information.
2. per gadget statistics: It provides the consolidated statistics for
each gadget running on the system.

TODO:
- why formatted memory and runtime aren't working?

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>