trace exec: add 'file' field for scripts
For scripts, the 'exepath' and 'file' fields are different. For example:
- exepath: "/bin/sh"
- file: "/tmp/script.sh"

'exepath' and 'file' are always set when the execution succeeds. When
the execution fails, it is best effort, depending on the error. For now,
they are set when the execution of a script fails because the schebang
is missing.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>