gadgets: trace_oomkill: Use gadget_process structure for from process.
For the target process, we do not use to use this structure, as we would only
fill 3 fields: the comm, PID and mntns_id.
For now, we cannot enrich on several mntns_id, so a warning is printed
indicating we have several fields of this type.
This warning will be fixed later, for now we ensure the enrichment is still done
on the target process.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>