trace_tcp: Add fds for connect and accept
The fields are documented in gadget.yaml:

- The 'fd' field is the fd where the operation happened on (connect,
  accept)
- The 'accept_fd' field is the FD of the new connection returned by
  accept, or -1 for errors cases or other events

The gadget uses maps to store context. Map cleanup is always done in a
tracepoint or a kprobe, avoiding kretprobe. This ensures it does not
miss a cleanup because the kretprobe overflows the maxactive.

TCP close events don't include a fd because this is not always
well-defined: a fd can be "renamed" with dup2() so it cannot be matched
with the fd from connect or accept events.

Signed-off-by: Burak Ok <burakok@microsoft.com>
Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>