trace exec gadgets: add exepath in case of failure
The trace exec gadgets have the following fields containing full paths:
- file: "/bin/script.sh"
- exepath: "/bin/sh"

They might be different when running a script, or they might be the same
when running a binary.

When the execution fails:
- the file field is set on a best effort basis, depending on the error
  (no changes here).
- the exepath field is set to the current program being executed (before
  this patch, it was left empty).

This new behaviour helps users to understand events when an execution
fails because a script does not have a shebang.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>