oci-service: Fix closing logic
The previous logic was causing issues when running the gadget failed
because the client was hanging forever. This is because if
`s.runtime.RunGadget` fails early the simple operator won't be
initialized and hence the done <- true operation will block as there
is not any goroutine reading from the channel.

This commit fixes that by closing the channel instead of writing to it.

Fixes: e689576d89bc ("pkg/gadget-service: update to support refactoring work for OCI gadgets")

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>