docs: Fix website building
The initial idea with 07141dcc5f2f ("docs/api: Improve Golang API documentation")
was to have a symlink from the docs to the examples, however it creates
a lot of issues while building the website as the whole building logic
only considers the /docs folder. This commit fixes that by doing the
other way around, moving the files to the /docs folder and creating
syminks from the examples to the docs.

Fixes: 07141dcc5f2f ("docs/api: Improve Golang API documentation")

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>