wasm: Remove malloc requirement for guest
Malloc() was required to be able to allocate and write data into the
guest memory. This requirement was a bit strict as:
- malloc() is exposed when using tinygo by a obscure undocumented way
- This isn't exposed when using go 1.24
- It can be challenging to expose on another programming languages

This commit removes this requirement. The main change is that the guest
now has to pass a preallocated buffer where the host will write the data.
It can be suboptimal for some cases when the guest doesn't know the
size, but it can help to reduce the allocations if the guest passes a
static buffer in multiple calls. (As it was done in the trace_exec and
trace_dns gadgets).

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>