pkg/gadget-context: sort operators by priority and name to be deterministic
Just taking the priority of an operator is not enough to get a deterministic order
of the operators. In case two operators have the same priority, the name should
also be considered for the run order.

Non-deterministic ordering can mix up for example fields in data sources, when multiple
nodes are running the same operators but in a different oder.

Signed-off-by: Michael Friese <mfriese@microsoft.com>