gadget params: rename print-*stack to collect-*stack
The --print-ustack and --print-kstack flags didn't control whether the
stacks are printed but whether they are collected (in ebpf code and in
the ustack operator). Hence it makes sense to rename them.

To print the stacks, users must either include the stack fields in the
column output with --fields, or use json/yaml output.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>