container-collection: reduce mem allocs for each event
When the podLabels and ownerReference fields are requested in each
event, it generates lots of allocations, which takes some cpu time
directly and more cpu time indirectly with the Go GC.

This patch reduces those allocations by keeping a cache of the
serialized podLabels string. Also, don't allocate the ownerReference
field twice.

Signed-off-by: Alban Crequy <albancrequy@linux.microsoft.com>