Makefile: Build ig for host architecture with go build instead of docker.
Building ig with docker can takes some time.
So, to build it for the host architecture, let's us go build instead.
Docker should continue to be used for different architecture, thus it is still
used when wanting to build ig when specifying the architecture like
ig-linux-amd64 or ig-linux-arm64.

Also, a new ig-on-docker targets was added to keep the old behavior in case
someone wants to build ig and does not have go installed.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>