gadget-service: improve logging of headless gadget instances
This changes the output format of gadget logs to JSON and adds the fields
"instanceID", "instanceName" and "gadget" to make them easier to digest for
existing postprocessing pipelines. This doesn't change the way the non-gadget
logging is handled, so that will still be in plain-text and should probably
be adjusted later on as well.

Signed-off-by: Michael Friese <mfriese@microsoft.com>