ustack: generic collect-ustack param and templates
Move GADGET_PARAM(collect_ustack) in the header file so it does not have
to be repeated in all gadgets' program.bpf.c.

Make it a well-known parameter so it does not have to be specified in
gadget.yaml.

Make use of templates so gadget.yaml does not have to repeat the
descriptions of the ustack fields.

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>