Document filter quoting
When specifying filters which contain special characters either on a
shell or in a YAML document, it's easy to run into quoting issues. For
example, specifying proc.comm~'^bash$' in a YAML document results in the
single quotes being passed down to IG which in turn causes unexpected
filtering behavior in user space.

Signed-off-by: Johanan Liebermann <jliebermann@microsoft.com>