operators/ebpf: Fix missing parameter default value for map-fetch-interval
The current logic relies on the client to set the value for the
parameters. This is done by kubectl-gadget, gadgetctl or ig by performing
a call to GetGadgetInfo() to get the list of parameters and their
default values. The issue here arises when using the headless mode, as
in this case, the call to GetGadgetInfo() is not done, so the value
for those parameters is missing if not explicitely set by the user in the
gadget manifest file. This behavior causes all gadgets that depend on the
"map-fetch-interval" param as it's set to 0.

The ideal solution for it should be something like
3c7f729e83ef ("pkg/gadget-service/api-helpers: Set default value for unset params"),
but it doesn't work in this case these parameters are returned by the
ExtraParams() method, hence they are not available when
InstantiateImageOperator() is called in pkg/operators/oci-handler/oci.go.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>