tests: Fix oci parameters
52d3310b7ad7 ("tests: Initialize operators with global parameters")
introduced logic to initialize the operators using the ones returned by
operators.GetDataOperators(), this isn't totally right as some operators,
as the oci-handler, are not present on operators.GetDataOperators().
This commit fixes it by using the right g.DataOperator variable that
contains the oci operator too.

There wasn't any issue so far until
bc6f56391fce ("oci: Fix handling of global and instance params") changed
the way parameters are initialized in the oci operator, causing a crash
in the test as Init() as never called, hence globalParms was never set.

This commit also uses the right variable for the verify-image param.

Fixes: 52d3310b7ad7 ("tests: Initialize operators with global parameters")

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>