operators/oci: Move params initialization back to InstantiateDataOperator()
bc6f56391fce ("oci: Fix handling of global and instance params") moved
the logic to set globalParams to the Init() function, even if this logic
should live on Init(), but we have so many places relying on this method
not being called, specially tests and examples. Let's move it back to
InstantiateDataOperator() to unblock the release.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>