pkg/kallsyms: re-enable test for kernel module symbol
Kallsyms test on symbol veth_init failed a lot, so it was disabled.

I believe this is due to using a function declared with "__init" because
the memory holding the code can be lazily reclaimed when the kernel
needs more memory,

This patch re-enables the test but uses a different symbol that does not
use "__init".

Fixes: d25e2295d1e6 ("pkg/kallsyms: disable failing test")

Signed-off-by: Alban Crequy <alban.crequy@gmail.com>