tlv experiments
Use TLV to optimize strings transfer from kernel to user space.

Issues:
- In order to use ring buffers, we need to allocate the buffer before,
which requires to know the size of the string, but to know the size of the
string we need to read the string. This solution allocates the buffer
with the maximum size of the string, otherwise a temporal buffer and one
extra copy will be needed.
- There are some strange verifier errors, likely because of bad castings
and/or overflows.